--- user-service/src/main/java/on/logistics/user_service/UserServiceApplication.java
@@ -1,13 +1,0 @@
-package on.logistics.user_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserServiceApplication.class, args);
-	}
-
-}
