--- user-service/src/main/java/on/logistics/user_service/global/Pageable.java
@@ -1,1 +1,0 @@
-package on.logistics.user_service.global;
