--- user-service/src/main/java/on/logistics/user_service/global/configuration/RedisConfig.java
@@ -1,5 +1,0 @@
-package on.logistics.user_service.global.configuration;
-
-public class RedisConfig {
-
-}
