--- user-service/src/main/java/on/logistics/user_service/global/enums/PageSize.java
@@ -1,22 +1,0 @@
-package on.logistics.user_service.global.enums;
-
-import java.util.Arrays;
-import lombok.Getter;
-
-@Getter
-public enum PageSize {
-    DEFAULT(10),
-    THIRTY(30),
-    FIFTY(50);
-
-    private final int size;
-
-    PageSize(int size) {
-        this.size = size;
-    }
-
-    public static boolean isValidSize(int size) {
-        return Arrays.stream(values())
-            .anyMatch(p -> p.getSize() == size);
-    }
-}
