--- user-service/src/main/java/on/logistics/user_service/global/exception/ExceptionCode.java
@@ -1,12 +1,0 @@
-package on.logistics.user_service.global.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionCode {
-
-    String name();
-
-    HttpStatus getHttpStatus();
-
-    String getMessage();
-}
