+++ user-service/src/main/java/on/logistics/userservice/UserServiceApplication.java
@@ -1,0 +1,13 @@
+package on.logistics.userservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserServiceApplication.class, args);
+	}
+
+}
