+++ user-service/src/main/java/on/logistics/userservice/application/dtos/CreateUserRequestDto.java
@@ -1,0 +1,5 @@
+package on.logistics.userservice.application.dtos;
+
+public class CreateUserRequestDto {
+
+}
