+++ user-service/src/main/java/on/logistics/userservice/application/dtos/UserCreateResponseDto.java
@@ -1,0 +1,7 @@
+package on.logistics.userservice.application.dtos;
+
+public record UserCreateResponseDto(
+
+) {
+
+}
