+++ user-service/src/main/java/on/logistics/userservice/global/configuration/JpaAuditingConfig.java
@@ -1,0 +1,10 @@
+package on.logistics.userservice.global.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+
+}
