+++ user-service/src/main/java/on/logistics/userservice/global/configuration/RedisConfig.java
@@ -1,0 +1,5 @@
+package on.logistics.userservice.global.configuration;
+
+public class RedisConfig {
+
+}
