+++ user-service/src/main/java/on/logistics/userservice/global/enums/PageSortBy.java
@@ -1,0 +1,22 @@
+package on.logistics.userservice.global.enums;
+
+import java.util.Arrays;
+import lombok.Getter;
+
+@Getter
+public enum PageSortBy {
+    CREATED_AT("createdAt"),
+    UPDATED_AT("updatedAt"),
+    ID("id");
+
+    private final String sortBy;
+
+    PageSortBy(String sortBy) {
+        this.sortBy = sortBy;
+    }
+
+    public static boolean isValidSortBy(String sortBy) {
+        return Arrays.stream(values())
+            .anyMatch(e -> e.getSortBy().equals(sortBy));
+    }
+}
