+++ user-service/src/main/java/on/logistics/userservice/global/exception/CustomException.java
@@ -1,0 +1,14 @@
+package on.logistics.userservice.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+
+    private final ExceptionCode exception;
+
+    public CustomException(ExceptionCode e) {
+        super(e.getMessage());
+        this.exception = e;
+    }
+}
