+++ user-service/src/main/java/on/logistics/userservice/infrastructure/querydsl/UserQueryRepository.java
@@ -1,0 +1,10 @@
+package on.logistics.userservice.infrastructure.querydsl;
+
+import on.logistics.userservice.domain.entity.User;
+import org.springframework.data.domain.Page;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserQueryRepository {
+
+}
