+++ user-service/src/main/java/on/logistics/userservice/infrastructure/querydsl/UserQueryRepositoryImpl.java
@@ -1,0 +1,15 @@
+package on.logistics.userservice.infrastructure.querydsl;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Repository;
+
+@Slf4j(topic = "UserQueryRepositoryImpl")
+@Repository
+@RequiredArgsConstructor
+public class UserQueryRepositoryImpl implements UserQueryRepository {
+
+    private final JPAQueryFactory queryFactory;
+
+}
