+++ user-service/src/main/java/on/logistics/userservice/presentation/dtos/UserCreateRequest.java
@@ -1,0 +1,8 @@
+package on.logistics.userservice.presentation.dtos;
+
+public record UserCreateRequest(
+    String nickname,
+    String slackEmail
+) {
+
+}
