+++ user-service/src/main/java/on/logistics/userservice/presentation/dtos/UserCreateResponse.java
@@ -1,0 +1,20 @@
+package on.logistics.userservice.presentation.dtos;
+
+import java.util.UUID;
+import on.logistics.userservice.domain.entity.User;
+
+public record UserCreateResponse(
+    UUID userId,
+    String nickname,
+    String slackEmail
+) {
+
+    public static UserCreateResponse from(User user) {
+        return new UserCreateResponse(
+            user.getId(),
+            user.getNickname(),
+            user.getSlackEmail()
+        );
+    }
+
+}
