+++ user-service/src/main/java/on/logistics/userservice/presentation/dtos/UserFindByIdResponse.java
@@ -1,0 +1,20 @@
+package on.logistics.userservice.presentation.dtos;
+
+
+import java.util.UUID;
+import on.logistics.userservice.domain.entity.User;
+
+public record UserFindByIdResponse(
+    UUID userId,
+    String nickname,
+    String slackEmail
+) {
+    public static UserFindByIdResponse from(User user) {
+        return new UserFindByIdResponse(
+            user.getId(),
+            user.getNickname(),
+            user.getSlackEmail()
+        );
+    }
+
+}
