+++ user-service/src/test/java/on/logistics/userservice/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package on.logistics.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
